Consolidate network code
