+++ Sources/tinys3/S3Client.swift
@@ -9,11 +9,11 @@ import FoundationNetworking
 import OSLog
 #endif
 
-public struct S3Client {
+public struct S3Client: RequestPerformer {
 
     private let credentials: AWSCredentials
     private let endpoint: S3Endpoint
-    private let urlSession: URLSession
+    let urlSession: URLSession
 
     public init(credentials: AWSCredentials, endpoint: S3Endpoint = .default, urlSession: URLSession = .shared) {
         self.credentials = credentials
@@ -55,7 +55,7 @@ public struct S3Client {
             credentials: self.credentials
         )
 
-        let response = try await perform(request: request).validate()
+        let response = try await perform(request).validate()
         return try S3ListResponse.from(response: response)
     }
 
@@ -97,35 +97,5 @@ public struct S3Client {
         )
 
         try await operation.start(progressCallback)
-    }
-
-    func perform(request: AWSRequest) async throws -> AWSResponse {
-        let (data, response) = try await perform(request: request.urlRequest)
-        return try AWSResponse(response: response, data: data).validate()
-    }
-
-    func perform(request: URLRequest) async throws -> (Data, HTTPURLResponse) {
-        return try await withCheckedThrowingContinuation { continuation -> Void in
-            let task = self.urlSession.dataTask(with: request) { data, response, networkError in
-                if let error = networkError {
-                    continuation.resume(throwing: error)
-                    return
-                }
-
-                if let data = data, let response = response as? HTTPURLResponse {
-                    continuation.resume(with: .success((data, response)))
-                    return
-                }
-
-                if let response = response as? HTTPURLResponse {
-                    continuation.resume(with: .success((Data(), response)))
-                    return
-                }
-
-                abort()
-            }
-
-            task.resume()
-        }
     }
 }
