+++ Sources/tinys3/model/S3Error.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum S3Error: Error {
+    case unknownHttpError
+}
